Add character counter for offer description
A new script (char-counter.js) has been added to implement a character counter for the offer description textarea in the 'create an offer' page. This counter shows the number of remaining characters allowed, automatically hiding when the area is empty and showing when characters are being input. The display style of the counter has also been adjusted within the CSS styling sheet.