Add a new Internship page and improve responsive design
Introduce a new Internship-page.html and enhance existing pages' responsive design. The company-profile.html now includes a collapsible sidebar for smaller screen sizes. Additionally, the threshold for hiding the sidebar in the edit-menu.js was slightly adjusted. Introduced new CSS styles in the style.css to manage the layout of the new internship page and other elements.