Refactor dropdown filter functionality
Introduced changes to the dropdown filter functionality in filter-fetch.js. This includes extracting the actual text of the dropdown value to display as the filter tag, instead of the code, and improving the data handling when fetching the cities via AJAX call. Also adjusted the way re-enabling of disabled options is being handled.