Implement search functionality in internship page but fuck it isn't implemented for internships yet :(
Refactored the loadInternships function in the script file, separating the URL into a variable for clearer code structure. A keypress function was added for the search bar to detect when the 'Enter' key is pressed. The updated functionality includes reloading the internships page with a filter based on the input if it is non-empty. If the input is empty, it will reload all internships.