Remove jQuery version 3.7.1 to replace it with something on the internet, added user page and research
The complete jQuery version 3.7.1 JavaScript file was removed from the project, as indicated by the diff showing the deletion of the entire content. This could be part of an upgrade process or a switch to another JavaScript library, or it might indicate that the project no longer needs jQuery. Any replacements, upgrades, or mitigation strategies implemented alongside this deletion should also be review-checked and tested for compatibility.