Display internship count and company rating in UI
The script now displays the number of internships a company offers and the average rating in the appropriate fields. Additionally, the code for generating stars according to the rating has been added. A placeholder has also been created for the total number of students working in the company. The 'api-handler.js' file has been updated to clear the paging container before populating it with new data.