Add student count and enhance filtering in load-companies.js
Added the number of students working in a company to the output of load-companies.js. Further improved the filtering of companies in company-profile.html by creating a dropdown menu pulling from an API, and wrote AJAX requests to fetch data for business sectors and cities for better user experience in choosing filtering options. Furthermore, the reset function was added to empty the companies-container.